requirements.txt
kaggle
pandas
apache-airflow-providers-amazon
apache-airflow[amazon]
jupyter
apache-airflow-providers-mysql
mysql-connector-python
